+++ app/controllers/metadata_presenter/change_answer_controller.rb
@@ -1,0 +1,8 @@
+module MetadataPresenter
+  class ChangeAnswerController < EngineController
+    def create
+      session[:return_to_check_you_answer] = true
+      redirect_to_page params[:url]
+    end
+  end
+end
