+++ app/controllers/metadata_presenter/engine_controller.rb
@@ -4,5 +4,22 @@ module MetadataPresenter
 
     helper MetadataPresenter::ApplicationHelper
     default_form_builder GOVUKDesignSystemFormBuilder::FormBuilder
+
+    def back_link
+      return if @page.blank?
+
+      @back_link ||= service.previous_page(current_page: @page)&.url
+    end
+    helper_method :back_link
+
+    private
+
+    def not_found
+      render template: 'errors/404', status: 404
+    end
+
+    def redirect_to_page(url)
+      redirect_to File.join(request.script_name, url)
+    end
   end
 end
