+++ app/controllers/metadata_presenter/submissions_controller.rb
@@ -1,0 +1,13 @@
+module MetadataPresenter
+  class SubmissionsController < EngineController
+    def create
+      @page = service.confirmation_page
+
+      if @page
+        redirect_to_page @page.url
+      else
+        not_found
+      end
+    end
+  end
+end
